Start sorting out the PSP directory selection strategy for a given Zen architecture version.
The on chip bootloader writes some sort of identification value into a specific register
right when it starts and reads that value later on to decide which L1 PSP directory to load from
flash and to select the proper off chip bootloader. The off chip bootloader also reads this value
multiple times. This requires us to set the value when execution is not started at the on chip
bootloader stage.